*** empty log message ***

+++ trunk/src/gmt_agc_io.c
@@ -1,4 +1,4 @@
-/*      $Id: gmt_agc_io.c,v 1.9 2006-10-24 20:01:09 pwessel Exp $
+/*      $Id: gmt_agc_io.c,v 1.10 2006-10-24 20:19:25 pwessel Exp $
  *
  * Based on original code from Robert Helie.  That code was hard-wired
  * in two applications (gmt2agcgrd.c and agc2gmtgrd.c) based on GMT 3.4.
@@ -65,7 +65,7 @@ int GMT_is_agc_grid (char *file)
 	if (nx <= 0) return (-1);
 	ny = GMT_get_n (y_min, y_max, y_inc, 0);
 	if (ny <= 0) return (-1);
-	predicted_size = irint (ceil (ny /ZBLOCKHEIGHT) * ceil (nx / ZBLOCKWIDTH)) * (ZBLOCKHEIGHT * ZBLOCKWIDTH + PREHEADSIZE + POSTHEADSIZE) * sizeof (float);
+	predicted_size = irint (ceil ((double)ny /ZBLOCKHEIGHT) * ceil ((double)nx / ZBLOCKWIDTH)) * (ZBLOCKHEIGHT * ZBLOCKWIDTH + PREHEADSIZE + POSTHEADSIZE) * sizeof (float);
 	if (predicted_size == buf.st_size) return (GMT_grd_format_decoder ("af"));
 	return (-1);
 }
